Get data from a form

This function connects to a specific form and fetches all answers. The responses are placed in a data.frame for easy use. For large data sets, toggle the incremental option, which while slow, will be better able to retrieve all responses. For retrieving responses only from a specific date on, or all responses after a certain submission, use the from_date and from_submission arguments. Forms that are anonymous and do not collect personal information do not record date, and as such the from_date will not work on those and an error will be thrown.

nettskjema_get_data(
  form_id,
  information = NULL,
  use_codebook = TRUE,
  as_is = FALSE,
  from_date = "",
  from_submission = "",
  incremental = FALSE,
  token_name = "NETTSKJEMA_API_TOKEN",
  ...
)

Arguments

form_id
integer. Number of the form to retrieve

information
character or list vector of additional answer information to add (One or more of "order", "option", "correct" "preselected"). If the list is named, the additional information suffixes will be changed to the list names. If NULLL (default) nothing is added. Currently not combinable with use_codebook = TRUE.

use_codebook
logical. Use codebook for retrieving answers (default TRUE). If you have not set up the codebook in nettskjema, then turning this off will return the data as is.

as_is
logical. Whether to return the raw content from nettskjema. Default is FALSE.

from_date
date. From which date on should data be fetched for

from_submission
integer. From which SubmissionId should data be collected from.

incremental
logical. False fetches all at once, TRUE fetches each submission individually. Slower but more stable for larger datasets.

token_name
character. Name to give the token, defaults to 'NETTSKJEMA_API_TOKEN'

...
arguments passed to GET

Value

tibble data.frame

Examples

if (FALSE) {

# Retrieve all data
data_110000 <- nettskjema_get_data(110000)
# Retrieve all data, and add answer order and option text
data_110000 <- nettskjema_get_data(110000, add)

# Retrieve all data after a certain date
data_110000 <- nettskjema_get_data(110000, from_date = "2019-01-01")

# Retrieve all submissions after a specific submission id
data_110000 <- nettskjema_get_data(110000, from_submission = "12000000")

# For large data
data_110000 <- nettskjema_get_data(110000, incremental = TRUE)

}